Split strategies into files
